fix(frontend): moved fetch environments to parent level
